Proper fix for the LRU IndexOutOfBoundsException
Previous fix for b/126427214 would result that the target process
being inserted gets falling out of the "Activities" section. A
proper fix should be in the remove LRU process part.

Also fixed an issue during the update of client activities:
when the sub proc should be sitting at the end of the group,
ineeded it's getting inserted to a position before end index
which could be pointing to a different process group, therefore
the end result would be "mixed" group.

Bug: 139765319
Test: atest CtsAppTestCases:ServiceTest\#testActivityServiceBindingLru
      manual: launch many apps and check the activity lru dump.

Change-Id: Ic56e38854837a07b8dabd1d19ba35eb1bfe4fd7a