Added logging in QsTileHost#onTuningChanged
Removed guards for most of the logging. Still guarded for tiles that are
added, as it goes through all the tiles every time something changes and
it's pretty verbose.

Also added log for destroyed tiles due to not being available.

These logs are not too frequent, happening only when a user is started
on the device and when the tiles are customized.

Test: adb logcat
Bug: 138848764
Change-Id: Ica203a6611d54ce22c9d426b0413e8b543ba41b3