Expose CdmaSmsCbProgramData as SystemApi
Only getOperation and getCategory are currently used, so we only expose
those parts of the functionality. Eventually if we move BearerData to
the CBR app we will also have to expose the constructor, at which point
we may also want to expose the other constants that the constructor
uses, but for now this is okay.

Bug: 135956699
Test: manual
Change-Id: I1d3ba564372845f9b8a470adf82e84df745c02a7
Merged-In: I1d3ba564372845f9b8a470adf82e84df745c02a7