Add wait flag to am switch-user command.
Previously, command returns right away. CTS tests that use it are forced to
loop waiting for the change in am get-current-user to see if the switch
completed. This method is unreliable because it returns the new user id before
UserController.mCurrentUser updated.

Adding "-w" to the switch command will make it wait for the switch to be
completed.

Fixes:141998718
Test: manual switch-user tests with -w flag
Change-Id: I41bbcff0c79677ba3bf6b58e1fc96cd13a681b33
Merged-In: I41bbcff0c79677ba3bf6b58e1fc96cd13a681b33